In progress, and Headers = 9 - Working!!!

+++ app/src/main/java/com/inflames1986/janweatherkotlin/model/repository/FactDTO.kt
@@ -30,5 +30,5 @@ data class FactDTO(
     @SerializedName("wind_gust")
     val windGust: Double,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
